# backend/docs/swagger.yaml
basePath: /
definitions:
  v1.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/v1.UserResponse'
    type: object
  v1.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  v1.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  v1.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  v1.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: Music streaming backend
  title: Chimera API
  version: "1.0"
paths:
  /v1/auth/login:
    post:
      consumes:
      - application/json
      operationId: login
      parameters:
      - description: Login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Login
      tags:
      - auth
  /v1/auth/register:
    post:
      consumes:
      - application/json
      operationId: register
      parameters:
      - description: Register
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Register user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: JWT from /v1/auth/login. Paste only the token, or "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"